refactor(dashboard): split fleet-telemetry-panel 897→410 lines
Split fleet-telemetry-panel.ts into UI-only panel + data utils:

- fleet-telemetry-utils.ts (514 lines): all pure functions — constants,
  types (FleetRow, FleetTelemetryState), normalizers, classifiers,
  formatters, warning builders, buildFleetRows, summaryCounts.
  Extracted verbatim via `sed -n '22,530p'` to avoid field-name drift.

- fleet-telemetry-panel.ts (410 lines): UI components only — SummaryCard,
  WarningBanner, PressureWatchlist, FleetComparisonTable,
  TelemetrySourcesPanel, FailureCategoryPanel, FleetTelemetryPanel (main).
  Re-exports buildFleetRows for test compat.

No behavior change. All 18 existing tests pass.
